// src/types/filters.ts
type ProjectFilters = {
  id_type_initiative?: number[];
  id_classification_management_area?: number[];
  id_clasification_meta_population?: number[];
  id_person_in_charge?: number[];
  id_university_body?: number[];
  id_user?: number[];
  id_region?: number[];
  id_university?: number[];
};

export default ProjectFilters;